SPRING works
need to set position limit to give impulse to ball

+++ Rainbow Pinball/ModulePlayer.h
@@ -15,6 +15,7 @@ public:
 
 public:
 
-	uint max_movement;
+	
+	float max_movement;
 
 };
